requirements.txt
Django==5.0.6
djangorestframework==3.15.2
channels==4.0.0
channels-redis==4.2.0
redis==5.0.4
PyJWT==2.9.0
daphne==4.1.2
django-cors-headers==4.4.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
